CalculiX 2.15 (new formula)
CalculiX is a software designed to solve field problems. It uses
the finite element method (FEM). With CalculiX Finite Element Models
can be built, calculated and post-processed. CalculiX is often used
together with FreeCAD's FEM workbench.

This patch is based on the following Formula:
https://github.com/brewsci/homebrew-science/blob/master/Formula/calculix-ccx.rb
in the now closed brewsci/homebrew-science tap repository.